requirements.txt
urllib3~=2.2.1  # https://github.com/urllib3/urllib3
python-dateutil~=2.9.0.post0 # https://github.com/dateutil/dateutil
Authlib~=1.3.0 # https://github.com/lepture/authlib
pyjwt~=2.8.0  # https://github.com/jpadilla/pyjwt
requests~=2.32.0 # https://github.com/psf/requests
typing-extensions~=4.11.0 # https://github.com/python/typing_extensions
frozendict~=2.4.3 # https://github.com/Marco-Sulla/python-frozendict
certifi~=2024.7.4 # https://github.com/certifi/python-certifi
# Core dependencies
django>=4.2.0,<5.0.0
python-dotenv>=1.0.0
requests>=2.32.2
cryptography>=41.0.0
pyjwt>=2.10.0
fastapi>=0.110.0
flask>=3.0.0
flask-session>=0.5.0
uvicorn>=0.25.0
python-dotenv>=1.0.0
itsdangerous>=2.2.0
nest-asyncio>=1.5.8
frozendict>=2.4.3,<3.0
python-dateutil>=2.9.0,<3.0
urllib3>=2.2.0,<3.0
Authlib>=1.3.0,<2.0
typing-extensions>=4.11.0,<5.0
certifi>=2024.2.2,<2025.0.0
pydantic>=2.0.0,<3.0.0

# Development dependencies
pytest>=7.4.0
pytest-django>=4.5.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-timeout>=2.2.0
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0
PyYAML>=6.0.1
twine>=4.0.2

# Other dependencies
httpx>=0.25.0,<0.29.0
